+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All Notable changes to `php-lxd` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+# unreleaed
+
+## Added
+  - Instance files are now project aware
+
 # [0.16.4]
 
 ## Added